Move osfs pkg into the jet core
